Move components  to shared for reuse
