+++ src/features/auth/pages/LoginPage.tsx
@@ -4,14 +4,16 @@ import { Link as RouterLink, useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
 import { yupResolver } from "@hookform/resolvers/yup";
 import { loginSchema } from "../schemas/loginSchema";
-import InputForm from "../components/InputForm";
+
 import AuthFormContainer from "../components/AuthFormContainer";
 import ButtonPrimary from "../../../shared/ui/components/ButtonPrimary";
 import { FormValues } from "../types";
-import HeaderForm from "../components/HeaderForm";
+
 import authService from "../services/AuthService";
 import { useAuthStore} from "../../../store/auth.store";
 import { showError, showSuccess } from "../../../shared/ui/toast";
+import HeaderForm from "../../../shared/ui/components/HeaderForm";
+import InputForm from "../../../shared/ui/components/InputForm";
 
 const LoginPage = () => {
   const navigate = useNavigate();
